Mac compile warnings

+++ src/common/macros.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#define vx_countof(arr) (sizeof(arr) / sizeof(arr[0]))
+#define vx_countof(arr) (int)(sizeof(arr) / sizeof(arr[0]))
 
 #define KB *(1024ull)
 #define MB *(1024ull * 1024ull)
